pyproject.toml
[project]
name = "resnet-unet-imagesegmentation"
version = "0.1.0"
description = "PyTorch ResNet50-UNet for semantic segmentation with CLI and notebook training"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "tqdm>=4.64.0",
    "matplotlib>=3.5.0",
    "tensorboard>=2.10.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
